Expose a few more host locations inside the container under /run/host
This is meant to alleviate some of the pain of not being able to modify
the list of bind mounts once a toolbox container has been created. For
some cases, especially where read-only access is enough, one can get
by with setting up symbolic links inside the toolbox container.

Based on an idea from Colin Walters.

https://github.com/debarshiray/toolbox/pull/264