Be forgiving of a missing /etc/profile.d/toolbox.sh in 'run'
It's common for people to run the toolbox script straight out of the
source tree without installing it system-wide. In such cases, it's
likely that /etc/profile.d/toolbox.sh would be absent on the host, and
as a result also absent from the toolbox container.

The welcome messages and the primary shell prompt (or PS1) are set
through /etc/profile.d/toolbox.sh, so not having it does degrade the
user experience, but it's probably not severe enough to fail the 'run'
command.

This should have been part of commit 0db54946b43e040 which split the
copying of /etc/profile.d/toolbox.sh into a container into two steps to
avoid using 'podman cp'. It already tried to handle the missing file
in the first step, but not in the second step.

It's also nice to at least make the user aware of the situation by
printing an error message.

https://github.com/debarshiray/toolbox/pull/278