Restructure bash-completion
The old bash completion script was quite messy so I gave it a bit of a
makeover and made some upgrades.

The script now doesn't offer global options (--verbose and such) if
they are already written. Command specific options are only completed
when commands are present in the line. And errors when listing images
and containers are redirected to null. This usually happens inside
containers.