Re-use a lower numbered file descriptor in 'reset'
POSIX only supports single digit file descriptors. Therefore, there's
value in being frugal about how we allocate them throughout the code.

The 'reset' command is very standalone and isolated from the other
code paths, because it's meant to be a last-ditch attempt to unbreak a
broken Podman installation. This can be exploited to re-use one of the
file descriptors that's used elsewhere in the code. In this case, file
descriptor number 4 is also used to control the spinner.

https://github.com/containers/toolbox/pull/305