Give access to /etc/machine-id from the host operating system
The machine ID is necessary to query the host operating system's
systemd journal, and currently toolbox containers have an empty
/etc/machine-id file.

Unlike /etc/resolv.conf, the machine ID is supposed to stay constant
once the host is booted. Therefore, it is safe to bind mount
/etc/machine-id from the host, as opposed to using a symbolic link;
because there's no chance of the file getting atomically updated on
the host and diverging from the bind mount due to being allocated a
new inode. Incidentally, this is also what Flatpak does.

A subsequent commit will use this to enable accessing the host's
systemd journal via journalctl(1) inside toolbox containers.

https://github.com/containers/toolbox/pull/327