cmd/run: Don't break GNU Readline's ctrl-p shortcut
Podman sets 'ctrl-p ctrl-q' as the default key sequence for detaching
a container. This breaks the ctrl-p shortcut that's equivalent to the
up arrow key in GNU Readline environments like Bash and Emacs.
Moreoever, toolbox containers aren't meant to be detached in the first
place.

Since Podman 1.8.1, it is now possible to unset the key sequence for
detaching [2, 3].

[0] https://tiswww.cwru.edu/php/chet/readline/readline.html#SEC15

[1] https://www.gnu.org/software/emacs/tour/

[2] Podman commit 7c623bd41ff3d534
    https://github.com/containers/podman/issues/4208

[3] Podman commit ebfd253fc658ffc9
    https://github.com/containers/podman/issues/5166

https://github.com/containers/toolbox/issues/394