Add optional container initialization hook
This adds support for a user-specified command to be run as the last
step in the container initialization process.  A new optional flag,
`--hook` is added to `init-container`, which is set based on another
new optional flag, `--init-hook`, to the `create` command.

If no hook is specified on the command-line, the image is inspected
for a `com.github.containers.toolbox.init` label, and the value is
used as the hook command if found.