cmd/root: Initialize globals on start-up
These values are generally useful, and reading any of them is a
potentially fallible operation. Since the absence of any of these
values is indicative of a broken environment, it's more convenient to
read them early on during start-up and abort, instead of trying to
handle the error later on.

https://github.com/containers/toolbox/pull/318