profile: Remove welcome prompt
This drops the hardcoded "Welcome to Fedora Silverblue"
which became wrong when "workstation" as added to the conditionals.

Now, the other change here is a followup to https://github.com/containers/toolbox/pull/221

My goal from the very start with rpm-ostree has been to
encourage using it in "image mode" by default.  But package layering
is 100% supported and certainly *allowed*.

So I object to use of the term "allow" in "allow package installation"
which implies it's not allowed.

We recommend containers.  But ultimately, it's *your* computer, not
ours.  So let's not use terms like "allow" or "immutable".

See also
https://blog.verbum.org/2019/12/23/starting-from-open-and-foss/

In preparation for moving prompts/displays like this downstream
into projects like Silverblue (and to encourage using toolbox
in other distributions) remove the prompt/welcome stuff entirely.