Add unique build id when building & specifiy binary name
When building a binary in Fedora the debug symbols and such are
automatically stripped and put in a *-devel package. This requires the
binary to have a unique build id[0].

Option '-o' in 'go build' makes the command create the final binary in a
specified path. When building for Fedora the lack of '/toolbox' caused
the binary to be called 'src'.

[0] https://fedoraproject.org/wiki/Releases/FeatureBuildId#Unique_build_ID

Signed-off-by: Ondřej Míchal <harrymichal@seznam.cz>