pkg/podman: Add and use package-wide error variables
Using Golang errors with dynamic strings is not very good because it's
very hard to check them. It is much easier to declare package-wide
errors based on which printed error messages are generated. This makes
out job to check for expected/unexpected errors much easier (even though
there's a bit more code for checking and acting upon the different
errors).