Update README
The README was in need of upgrading. The structure of the README is a
bit different now:

- Few words about Toolbox
- Basic usage (few commands)
- More info about Toolbox
- Link to Documentation
- How to install
- How to build from source
- Distro support
- Image requirements
- Goals and Use Cases

Mainly the introduction of Toolbox received a major facelift. There's a
also a new paragraph about Toolbox not being a secure isolation tool.

The installation section now also mentions Arch Linux where AFAIK some
people are using Toolbox quite successfully.

The build from source instructions were also upgraded and some badges
(resp. shields) were added for quick access to the package info and
tests dashboard (Zuul[0] recently added support for dynamic badges. When
I figure out how to access it, we could put it into the README).

The distro support section also was extended to clarify more how Toolbox
operates on different OSs and with different OCI images.

I tried to update wording in some places (e.g., dropped the word
'unprivileged' because users often thought it means 'isolated' or
'sandboxed' which is not true).

[0] https://zuul-ci.org/
[1] https://softwarefactory-project.io