cmd/run: Run a login shell when falling back to Bash during 'enter'
'toolbox enter' attempts to run the user's preferred shell as a login
shell by passing it the '-l' argument. However, when the user's
preferred shell is missing from the toolbox container and it falls
back to Bash, then it was no longer a login shell because the '-l'
argument went missing.

https://github.com/containers/toolbox/pull/588