cmd/initContainer: Do create a symlink if the containerPath was absent
The redirectPath function used to error out when handling directories,
if the path inside the container was initially absent. There's no real
reason for this, and some containers failed to start because the
/media directory was absent from them. This can happen as a
consequence of Fedora's filesystem RPM failing 'dnf update'
transactions inside containers:
  $ toolbox enter f33-to-upgrade
  $ dnf --assumeyes --enablerepo=fedora --disablerepo=rawhide update
  <...>
  Failed:
    filesystem-3.14-2.fc32.x86_64
    filesystem-3.14-3.fc33.x86_64

  Error: Transaction failed

Therefore, it's better to not worry about the path being initially
absent, and be more forgiving and robust.

https://github.com/containers/toolbox/issues/539