cmd/initContainer: Explicitly set-up X11 sockets in /tmp
Toolbox until now relied on the existence of an abstract unix socket
(more in unix(7)) for the communication with X server. In version 3.38
GNOME Shell stopped using this socket and instead uses the "standard"
approach of using the one found in /tmp. And since Toolbox mounts a
tmpfs to /tmp instead of the hosts, /tmp in a toolbox is empty.

With this the crucial paths /run/host/tmp are symlinked to /tmp,
enabling (again) running of GUI applications from a toolbox.