cmd/initContainer: Add types for mounts and symlinks
There might be cases when a "symlink/mount candidate" need to be created
and for such occasions it's better to have a type than having to define
the struct again and again.

I'm also thinking of connecting the redirectPath() and mountBind()
functions to the types. That is currently just an idea, though.