cmd/initContainer: Only remove passwords when needed
When looking into logs of toolboxes using 'podman logs', one can notice
that every "startup" log mentions removing the password for the user and
root. These lines could be considered as bloat because what they're
saying is actually not happening because appart from the first run the
password are usually already gone.

This makes the removal of password for the user and root conditional
based on the output of 'passwd --status <username>' where value "NP"
(meaning "no password") will be considered the only value that does not
trigger the removal. Any other value will trigger the removal.