src/go-build-wrapper: Support setting build options via env var
Toolbox uses a build wrapper that is used by Meson becase meson does not
support natively Go. That proved to be troublesome when we tried to
build Toolbox in Fedora's buildsystem that sets a number of CFLAGS,
LDFLAGS and Go specific flags. There was no way to pass those to the
wrapper, hence we hardcoded the most important ones using patches. With
this the patches can be dropped.

To pass an option to the 'go build' command, set GOBUILDFLAGS
environmental variable and export it.

https://github.com/containers/toolbox/pull/583