cmd/run: Fill out the remaining fields in the OSC 777 escape sequence
A subsequent commit will add another argument to the OSC 777 escape
sequence to specify the user ID that toolbox(1) is running as. Unlike
the NAME and RUNTIME arguments, the UID argument won't be optional when
popping. Therefore it's nicer to spell everything out for the sake of
completeness.

This change relies on VTE reading in all the remaining arguments when
popping. Otherwise, the argument(s) might get spuriously displayed in
the terminal.

https://github.com/containers/toolbox/issues/267